refactor boton amigos
