+++ webapp/src/components/Pages/WelcomePage/WelcomePage.tsx
@@ -3,23 +3,12 @@ import WelcomeText from "./subcomponents/WelcomeText";
 import WelcomeSolid from "./subcomponents/WelcomeSolid";
 import {
   OSMap,
-  ShowMarkersFulfilledPromise,
 } from "../../Map/OSMap";
 import "aos/dist/aos.css";
 import { useSession } from "@inrupt/solid-ui-react";
-import { getFriendsFromPod, readFromDataSet, readFromFriendDataSet } from "../../Solid/ReadFromPod";
-import { useState } from "react";
 
 export default function WelcomePage() {
   const { session } = useSession();
-  const [markers, setMarkers] = useState(readFromDataSet());
-  
-  function readFromFriend(url: string){
-    setMarkers(readFromFriendDataSet(url));
-    markers.then((array: any) => {
-      ShowMarkersFulfilledPromise(array);
-    });
-  }
 
   return (
     <div className="welcome_page">
@@ -42,21 +31,6 @@ export default function WelcomePage() {
         </>
       ) : (
         <div className="map-container">
-          <button
-            onClick={() => {
-              try {
-                getFriendsFromPod().then((friends: any) => {
-                  friends.forEach((friend: any) => {
-                    readFromFriend(friend);
-                  });
-                });
-              } catch (error) {
-                console.log(error);
-              }
-            }}
-          >
-            Amigos
-          </button>
           <div className="map" data-aos="fade-down">
             <OSMap />
           </div>
